Safari-609.1.11.1/LayoutTests/http/tests/workers/self-readonly-expected.txt

CONSOLE MESSAGE: line 9: NaN
Test passes if "NaN" is returned by the worker.
